requirements.txt
python>=3.9
requests
jupyter
pytz
gdal>=3.0.0
matplotlib
numpy
scipy
pandas
tqdm
pylama
yapf
pytest
scikit-image
scikit-learn
pytorch
torchvision
pytorch-lightning
kornia